requirements.txt
Django==2.0.8
psycopg2==2.7.3.2
pyyaml==5.1
pytest==4.4.0
pytest-django==3.4.8
mixer==6.1.3
cerberus==1.2